Extract paint to paint.js and inline properties.js in index.js
